requirements.txt
speechrecognition
pyaudio
mss
pillow
torch
torchvision
opencv-python
numpy
pyttsx3
requests
ultralytics
open_clip_torch
openai
PyQt5
google-cloud-vision